+++ src/lib/string.c
@@ -4,7 +4,7 @@
 void *memset(void *dest, int c, uint64_t n){
     uint8_t *p = (uint8_t *)dest;
 
-    for(size_t i = 0; i < n; i++){
+    for(uint64_t i = 0; i < n; i++){
         p[i] = (uint8_t)c;
     }
 
@@ -15,7 +15,7 @@ void *memcpy(void *dest, const void *src, uint64_t n){
     uint8_t *pdest = (uint8_t *)dest;
     const uint8_t *psrc = (const uint8_t *)src;
 
-    for(size_t i = 0; i < n; i++){
+    for(uint64_t i = 0; i < n; i++){
         pdest[i] = psrc[i];
     }
 
@@ -28,11 +28,11 @@ void *memmove(void *dest, const void *src, uint64_t n) {
     const uint8_t *psrc = (const uint8_t *)src;
 
     if(src > dest){
-        for (size_t i = 0; i < n; i++) {
+        for (uint64_t i = 0; i < n; i++) {
             pdest[i] = psrc[i];
         }
     }else if(src < dest){
-        for (size_t i = n; i > 0; i--) {
+        for (uint64_t i = n; i > 0; i--) {
             pdest[i-1] = psrc[i-1];
         }
     }
@@ -44,7 +44,7 @@ int memcmp(const void *s1, const void *s2, uint64_t n){
     const uint8_t *p1 = (const uint8_t *)s1;
     const uint8_t *p2 = (const uint8_t *)s2;
 
-    for(size_t i = 0; i < n; i++){
+    for(uint64_t i = 0; i < n; i++){
         if(p1[i] != p2[i]){
             return p1[i] < p2[i] ? -1 : 1;
         }
